+++ x-objects/test/x/objects/list_test.rb
@@ -64,7 +64,7 @@ module X
     end
 
     def test_find_all_is_not_supported
-      error = assert_raises(NotImplementedError) { List.find_all([1, 2], client: @client) }
+      error = assert_raises(UnsupportedOperation) { List.find_all([1, 2], client: @client) }
 
       assert_equal "X::List cannot be fetched in batches; find 2 lists one at a time", error.message
       assert_empty @client.requests